requirements.txt
django-cms
feedparser
